[Marcus] - configurações de teste e setup

+++ database/migrations/2025_06_18_091149_create_favorite_products_table.php
@@ -8,20 +8,20 @@ return new class extends Migration
 {
     public function up(): void
     {
-        Schema::create('client.favorite_products', function (Blueprint $table) {
-            $table->id();
-            $table->unsignedBigInteger('client_id');
-            $table->unsignedBigInteger('product_id');
-            $table->timestamps();
-            $table->unique(['client_id', 'product_id']);
-
-
-            $table->foreign('client_id')
-                  ->references('id')
-                  ->on('client.clients')
-                  ->onDelete('cascade');
-        });
-
+        if (!Schema::hasTable('client.favorite_products')) {
+            Schema::create('client.favorite_products', function (Blueprint $table) {
+                $table->id();
+                $table->unsignedBigInteger('client_id');
+                $table->unsignedBigInteger('product_id');
+                $table->timestamps();
+                $table->unique(['client_id', 'product_id']);
+                /** Relacionamentos */
+                $table->foreign('client_id')
+                    ->references('id')
+                    ->on('client.clients')
+                    ->onDelete('cascade');
+            });
+        }
     }
 
     public function down(): void
